refactor: User 엔티티 경로 수정
- User 엔티티의 패키지 경로가 `com.springboot.monew.users.entity`에서 `com.springboot.monew.user.entity`로 변경됨에 따라,
- 테스트 코드에서 해당 엔티티를 import하는 경로를 수정했습니다.
