+++ src/test/java/com/springboot/monew/testdata/LoadTestDataRunner.java
@@ -21,10 +21,9 @@ import com.springboot.monew.testdata.entity.NewsArticleGenerator;
 import com.springboot.monew.testdata.entity.NotificationGenerator;
 import com.springboot.monew.testdata.entity.SubscriptionGenerator;
 import com.springboot.monew.testdata.entity.UserGenerator;
-import com.springboot.monew.users.entity.User;
+import com.springboot.monew.user.entity.User;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
