+++ src/test/java/com/springboot/monew/testdata/entity/NotificationGenerator.java
@@ -6,7 +6,7 @@ import com.springboot.monew.comment.entity.CommentLike;
 import com.springboot.monew.interest.entity.Interest;
 import com.springboot.monew.notification.entity.Notification;
 import com.springboot.monew.notification.entity.ResourceType;
-import com.springboot.monew.users.entity.User;
+import com.springboot.monew.user.entity.User;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.sql.Timestamp;
