+++ src/test/java/com/springboot/monew/testdata/entity/SubscriptionGenerator.java
@@ -4,7 +4,7 @@ import static org.instancio.Select.field;
 
 import com.springboot.monew.interest.entity.Interest;
 import com.springboot.monew.interest.entity.Subscription;
-import com.springboot.monew.users.entity.User;
+import com.springboot.monew.user.entity.User;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.sql.Timestamp;
