+++ src/test/java/com/springboot/monew/testdata/entity/UserGenerator.java
@@ -2,7 +2,7 @@ package com.springboot.monew.testdata.entity;
 
 import static org.instancio.Select.field;
 
-import com.springboot.monew.users.entity.User;
+import com.springboot.monew.user.entity.User;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.sql.Timestamp;
